+++ R/check.r
@@ -17,6 +17,9 @@ check <- function(x, ...) UseMethod("check")
 ##' @param contact.age.column the column in the \code{contacts} data frame containing contacts' age
 ##' @param ... ignored
 ##' @return invisibly returns a character vector of the relevant columns
+##' @examples
+##' data(polymod)
+##' check(polymod)
 ##' @export
 check.survey <- function(x, columns=FALSE, quiet=FALSE, error=FALSE, id.column="part_id", participant.age.column="part_age", country.column="country", year.column="year", contact.age.column="cnt_age", ...)
 {
@@ -32,7 +35,7 @@ check.survey <- function(x, columns=FALSE, quiet=FALSE, error=FALSE, id.column="
         if (!(id.column %in% colnames(x$participants) &&
               id.column %in% colnames(x$contacts)))
         {
-            error_func("Error: columns$id '", columns$id, "' does not exist in both the ",
+            error_func("Error: id.columns '", id.column, "' does not exist in both the ",
                  "participants and contacts data frames")
             success <- FALSE
         }