+++ NEWS.md
@@ -1,3 +1,10 @@
+# socialmixr 0.1.1
+
+* fix: dependency on R >= 3.4.0
+* updated reference handling if a survey is not pulled from Zenodo
+* the 'survey' function can now be used to create a 'survey' object from data frames
+* a few small patches and fixes
+
 # socialmixr 0.1.0
 
 * initial release